requirements.txt
sounddevice>=0.4.6
numpy>=1.24
openai-whisper>=20231117
pytest>=8.0
pyannote.audio>=3.1
